+++ components/Filter/index.tsx
@@ -1,31 +1,10 @@
 "use client";
-import { useEffect } from "react";
-import { useFormContext } from "react-hook-form";
 import axios from "axios";
 import { Button, Flex } from "antd";
 import Preferences from "./Preferences";
-import { getCompanies } from "@/app/actions";
+import Scrape from "./Scrape";
 
 export default async function Filter() {
-  useEffect(() => {
-    async function fetchCompanies() {
-      const companies = await getCompanies();
-      console.log("companies", companies);
-    }
-
-    fetchCompanies();
-  }, []);
-
-  const handleScrape = async () => {
-    try {
-      const response = await axios("/api/scrape");
-      const data = await response?.data;
-      console.log("Search API Response:", data);
-    } catch (error) {
-      console.error("Error fetching data:", error);
-    }
-  };
-
   const handleEnrich = async () => {
     try {
       const response = await axios("/api/enrich");
@@ -45,9 +24,7 @@ export default async function Filter() {
       }}
     >
       <Flex gap={20}>
-        <Button type="primary" onClick={handleScrape}>
-          Scrape
-        </Button>
+        <Scrape />
         <Button type="primary" onClick={handleEnrich}>
           Enrich
         </Button>
